requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9
Werkzeug==2.3.8
python-dotenv==1.0.0
